Removed unnecessary bug-inducing virtual tile shrinking code.
The bug was that the function which shrinks virtual tiles was set up
not to shrink the virtual tile along a dimension for which it
osculates a tile thinner than the virtual tile along that
dimension. That doesn't make sense and is inappropriate, it was
intended to avoid shrinking tiles by more than their width along a
dimension but that is accomplished correctly by
propagate_tile_perturbation.

Also, I added logging statements to notify when a virtual tile is
shrunk and when virtual tiles exist which cannot be shrunk.